Split views.py into 5 separate files, by functionality #119
