+++ SA3/qa_web/views/__init__.py
@@ -1,0 +1,5 @@
+from .add_content import *
+from .auth import *
+from .home import *
+from .profile import *
+from .question_operations import *
