+++ SA3/qa_web/views/home.py
@@ -1,0 +1,9 @@
+from django.shortcuts import render
+
+def homepage(request):
+    """Displays the website's home page.
+
+    :param request: Request data provided by the WSGI
+    :return: Rendered template displaying the home page
+    """
+    return render(request, "qa_web/home.html")
